Prevent unnecessary re-renders for the `CountdownContainer` component
The `CountdownContainer` contains the `Countdown` component and the `Countdown` component handles to countdown logic
