+++ src/components/countdownContainer/CountdownContainer.jsx
@@ -1,20 +1,11 @@
 import { useEffect, useState } from "react";
 
 import "./countdown-container.css";
+import Countdown from "./Countdown";
 
 function CountdownContainer({ title, date, setView }) {
   const countdownValue = new Date(date).getTime();
   const now = new Date().getTime();
-  const [difference, setDifference] = useState(countdownValue - now);
-
-  useEffect(() => {
-    const myInterval = setInterval(() => {
-      setDifference((prevDeffirence) => prevDeffirence - 1000);
-    }, 1000);
-
-    // clean up the inetervall when the coundown ends or when the component unmounts
-    return () => clearInterval(myInterval);
-  }, []);
 
   // hide the countdown and show the input
   function handleReset() {
@@ -23,34 +14,12 @@ function CountdownContainer({ title, date, setView }) {
     localStorage.removeItem("countdown");
   }
 
-  // update view when countdown reaches 0
-  useEffect(() => {
-    if (difference < 0) {
-      setView("complete");
-    }
-  }, [difference]);
-
   return (
     <div className="countdown-container">
       <h1>{title}</h1>
-      <ul>
-        <li>
-          <span>{Math.floor(difference / (1000 * 60 * 60 * 24))}</span>
-          Days
-        </li>
-        <li>
-          <span>{Math.floor(difference / (1000 * 60 * 60)) % 24}</span>
-          Hours
-        </li>
-        <li>
-          <span>{Math.floor(difference / (1000 * 60)) % 60}</span>
-          Minutes
-        </li>
-        <li>
-          <span>{Math.floor(difference / 1000) % 60}</span>
-          Seconds
-        </li>
-      </ul>
+
+      <Countdown countdownValue={countdownValue} now={now} setView={setView} />
+
       <button className="reset-btn" onClick={handleReset}>
         Reset
       </button>
